backend/requirements.txt
django
django-cors-headers
djangorestframework-simplejwt
rest_framework_simplejwt
psycopg2
